fix(dtt-121): add pre-check for latest files
The problem was that the directory created was done so using the UUID in
the url. Then we were doing `/v0/data/uuid?latest`, so the directory
name was the always the old UUID even if there was a new version.

Using the /files/versions endpoint we can tell the latest version of a
file before going to download it. Replacing the old list of UUIDs with
the new list of UUIDs with the newer versions in it. After that the download
process continues as normal.