feat(DTT-154): Remove pyinstaller workaround
Prior to the recent Python3 upgrade we used an older version of the
requests library. This library came with a `cacert.pem` file needed to
make https connections. This file was separated out into it's own
library called `certifi`, and PyInstaller added a hook to include this
file on build time. The original workaround assumed the cacert.pem file
existed in the requests folder of the temp directory made by PyInstaller.

The workaround is no longer needed because of the PyInstaller hook.

Other changes in this commit revolve around Python3 progressbar2 fixes.

- `e.message` is no longer needed in Python3.
- progress bar library was changed from progressbar to progressbar2 and
the behavior changed slightly.
- Reorganized imports in upload/client.py to pep8 recommendations.
- Added default parameter of True on verify request for download.
- add ProgressBar initial_value param in constructor
- write initial byte as byte and not string