DEV-401 fix race condition in multipart upload and refactor test
The counter variable being used in Manager Namespace() member variable
was being passed into each process, and the update was not being
synchronized among the processes, causing the completed variable to be
incorrect sporadically. There doesn't seem to be any need for the
manager namepsace, as we already wait for all futures to return before
iterating through the results.