DEV-401 Fix multipart upload progress logging (#269)
DEV-401: Fix multipart upload progress logging

When part fails for multipart upload, progressbar is no longer updated.
The additional retry of the whole multipart upload is redundant and leads to misleading logging,
so it is removed.

The counter variable being used in Manager Namespace() member variable
was being passed into each process, and the update was not being
synchronized among the processes, causing the completed variable to be
incorrect sporadically. There doesn't seem to be any need for the
manager namepsace, as we already wait for all futures to return before
iterating through the results.